requirements.dev.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
fastapi==0.70.1
filelock==3.4.0
fugashi==1.1.1
h11==0.12.0
huggingface-hub==0.2.1
idna==3.3
ipadic==1.0.0
joblib==1.1.0
mecab-python3==1.0.4
numpy==1.21.5
packaging==21.3
pydantic==1.8.2
pyparsing==3.0.6
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
sacremoses==0.0.46
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tokenizers==0.10.3
torch==1.10.1
tqdm==4.62.3
transformers==4.14.1
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
